[grid] Lift Active up to a top-level class
